+++ libdl/elf_hash.h
@@ -1,5 +1,5 @@
-static unsigned long elf_hash(const unsigned char *name) {
-  unsigned long h=0, g;
+static unsigned int elf_hash(const unsigned char *name) {
+  unsigned int h=0, g;
 
   while (*name) {
     h = (h<<4) + *(name++);
